requirements.txt
python-postmark==0.7.0
python-dotenv>=1.2.1
requests>=2.32.5
